recurso listagem de produtos do estoque retornando 200

+++ src/main/kotlin/com/act/spa/dto/response/ProdutoModelResponseDTO.kt
@@ -4,6 +4,6 @@ data class ProdutoModelResponseDTO(
     var codigo: String,
     var nome: String,
     var descricao: String,
-    var preco: Int,
+    var preco: Double,
     var quantidade: Int
 )
